+++ library/LibUtilities/TimeIntegration/CNABTimeIntegrationScheme.h
@@ -67,10 +67,11 @@ public:
         m_integration_phases[1] = TimeIntegrationAlgorithmGLMSharedPtr(
             new TimeIntegrationAlgorithmGLM(this));
 
+        // IMEX dirk 2 2 2
         IMEXdirkTimeIntegrationScheme::SetupSchemeData(m_integration_phases[0],
-                                                       2, {2, 2}); // dirk 2 2 2
-        CNABTimeIntegrationScheme::SetupSchemeData(
-            m_integration_phases[1]); // CNAB
+                                                       2, {2, 2});
+        // CNAB
+        CNABTimeIntegrationScheme::SetupSchemeData(m_integration_phases[1]);
     }
 
     virtual ~CNABTimeIntegrationScheme()
@@ -119,6 +120,7 @@ public:
         phase->m_V =
             Array<TwoD, NekDouble>(phase->m_numsteps, phase->m_numsteps, 0.0);
 
+        // Coefficients
         phase->m_A[0][0][0] = 1.0 / 2.0;
         phase->m_B[0][0][0] = 1.0 / 2.0;
         phase->m_B[0][1][0] = 1.0;